Authorization with VK, Friends Screen ui + logic return to add ModalDrawerSheet when Profile, Chat screens are ready

+++ core/data/src/main/java/com/laru/data/repository/ChatRepository.kt
@@ -3,5 +3,5 @@ package com.laru.data.repository
 
 interface ChatRepository {
 
-    suspend fun initializeMessages(chatId: Int)
+    suspend fun initializeMessages(chatId: Long)
 }
