+++ core/data/src/main/java/com/laru/data/repository/ChatRepositoryImpl.kt
@@ -5,7 +5,7 @@ import javax.inject.Inject
 class ChatRepositoryImpl @Inject constructor(
 
 ): ChatRepository {
-    override suspend fun initializeMessages(chatId: Int) {
+    override suspend fun initializeMessages(chatId: Long) {
 
     }
 }
